+++ cms/src/layouts/RootLayout.tsx
@@ -2,9 +2,13 @@ import { Outlet } from "@tanstack/react-router";
 
 import Header from "../components/header/Header";
 import Footer from "../components/footer/Footer";
+import { useAuth } from "../hooks/useAuth";
 
 const RootLayout = () => {
-  console.log("RootLayout running..");
+  console.log("[RootLayout] rendering...");
+  const auth = useAuth();
+
+  console.log("auth:", auth);
 
   return (
     <>