+++ cms/src/pages/dashboard/Dashboard.tsx
@@ -1,9 +1,13 @@
 import React from "react";
 
 import { useAuth } from "../../hooks/useAuth";
+import { useRouteContext } from "@tanstack/react-router";
 const Dashboard: React.FC = () => {
   console.group("Dashboard component rendering...");
+  const context = useRouteContext({ from: "/_protected" });
   const { accessToken } = useAuth();
+  console.log("context:", context);
+  console.groupEnd();
 
   return (
     <section>